Find secondary guests by BRN (WORKING!)
