+++ src/main/java/cs127/springappbe/Controller/BRNController.java
@@ -4,6 +4,7 @@ import cs127.springappbe.Entities.BRN;
 import cs127.springappbe.Entities.Request.AddBookingRequest;
 import cs127.springappbe.Entities.SecondaryGuest;
 import cs127.springappbe.Service.BRNService;
+import cs127.springappbe.Service.SGuestService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.*;
@@ -18,6 +19,9 @@ public class BRNController {
     @Qualifier("BRNServiceImpl")
     private final BRNService brnService;
 
+    @Qualifier("SGuestServiceImpl")
+    private final SGuestService sGuestService;
+
     @PostMapping(path="/add")
     public @ResponseBody BRN addBooking(@RequestBody AddBookingRequest bookingToAdd){
         return brnService.addBooking(bookingToAdd);
@@ -25,6 +29,6 @@ public class BRNController {
 
     @GetMapping(path="/secondaryguests")
     public List<SecondaryGuest> findSecondaryGuests(@RequestParam(value="BRNID") Long BRNID){
-        return brnService.findSecondaryGuests(BRNID);
+        return sGuestService.findSecondaryGuests(BRNID);
     }
 }
