+++ src/main/java/cs127/springappbe/Repository/BRNRepository.java
@@ -11,13 +11,13 @@ import java.util.List;
 @Repository
 public interface BRNRepository extends JpaRepository<BRN, Long> {
 
-    @Query(
-            value = "SELECT SG.* FROM BRN " +
-                    "FULL JOIN SPRINGAPPUSER.BRN_SGUEST BS on BRN.BRN_ID = BS.BRN_ID " +
-                    "FULL JOIN SPRINGAPPUSER.SECONDARY_GUEST SG on BS.SGUEST_ID = SG.SGUEST_ID " +
-                    "WHERE BS.BRN_ID=?1",
-            nativeQuery = true
-    )
-    List<SecondaryGuest> findSecondaryGuests(Long BRNID);
+//    @Query(
+//            value = "SELECT SG.* FROM BRN " +
+//                    "FULL JOIN SPRINGAPPUSER.BRN_SGUEST BS on BRN.BRN_ID = BS.BRN_ID " +
+//                    "FULL JOIN SPRINGAPPUSER.SECONDARY_GUEST SG on BS.SGUEST_ID = SG.SGUEST_ID " +
+//                    "WHERE BS.BRN_ID=?1",
+//            nativeQuery = true
+//    )
+//    List<SecondaryGuest> findSecondaryGuests(Long BRNID);
 
 }
